Update and rename README.md to Autosys_1.txt
Autosys is one of the most popular job scheduling tools used in the software development lifecycle process. It is a task scheduler created by using a batch program or UNIX script. To create/run an Autosys jobs, a user needs to have either a LINUX or Windows machine. Here are some of heavily used Autosys Commands.